Disable get_multilibs on Darwin.
All values "successfully" determined on Darwin are useless, and
attempts to determine it may involve invoking the compiler with
the -dumpspecs option, which fails miserably on non-gcc compilers
(e.g., clang), and even with gcc bloats the logfile with its
huge output.

Setting multitop to an empty string short-circuits the processing.

TESTED:
Using libffi tests as an example, ran on OSX 10.4-10.5 ppc, 10.4-10.6
i386, and 10.6-10.15 x86_64.  All tests behaved as intended, with
get_multilibs successfully returning empty string in all cases.
No change in behavior on Ubuntu 14.04, CentOS 7, or Fedora 25.
Pre-broken on *BSD.